+++ Lib/fontbakery/profiles/typenetwork.py
@@ -91,7 +91,6 @@ PROFILE = {
             "googlefonts:varfont/bold_wght_coord",
             "googlefonts:varfont/duplicate_instance_names",
             #
-            "smart_dropout",
             "varfont/consistent_axes",
         ],
         "Outline Checks": [